method call and reference

+++ src/reference/FacebookCustomer.java
@@ -1,18 +1,6 @@
 package reference;
-// Blue print of facebook customer;s profile
+
 public class FacebookCustomer {
 	String name;
-	byte age; 
+	int age;
 }
-// default values
-// all object data type = null
-// all numeric data type = 0
-// all decimal data type =0.0
-// all boolean data type = false
-// all char data type = 'u/0'
-
-// class is a blue print
-// object is realtime instance
-
-// fb class/blue print : every customer needs to have a name and age
-// fb object/real instance : here is a customer whose name is john and age 21
